+++ pymeasure/experiment/parameters.py
@@ -21,23 +21,36 @@
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 # THE SOFTWARE.
 #
+from warnings import warn
 
 
-class Parameter:
-    """ Encapsulates the information for an experiment parameter
-    with information about the name, and units if supplied.
+class InputField:
+    """ Encapsulates the information for an input of an experiment with information about the name.
 
-    :var value: The value of the parameter
+    :class:`.InputField` serves as a parent-class to :class:`.Parameter` and :class:`.Metadata` (and
+    their respective subclasses). These classes differ in whether (and if yes, when) their value is
+    writen to the data-file, and whether the value is applied when getting parameters from a file.
 
-    :param name: The parameter name
+    - :class:`.InputField` is not written to file, nor applied when getting parameters from a file.
+    - :class:`.Parameter` is written to the "Parameters" section of the file when the experiment is
+      queued.
+    - :class:`.Metadata` is writen to file after the
+      :py:meth:`~pymeasure.experiment.procedure.Procedure.startup` of a
+      :class:`~pymeasure.experiment.procedure.Procedure` is run and allows for storing information
+      that is to be retrieved upon the start of the procedure (such as the starting time or specific
+      instrument attributes or settings).
+
+    :var value: The value of the :class:`.InputField`
+
+    :param name: The :class:`.InputField` name
     :param default: The default value
-    :param ui_class: A Qt class to use for the UI of this parameter
-    :param group_by: Defines the Parameter(s) that controls the visibility
-        of the associated input; can be a string containing the Parameter
-        name, a list of strings with multiple Parameter names, or a dict
-        containing {"Parameter name": condition} pairs.
-    :param group_condition: The condition for the group_by Parameter
-        that controls the visibility of this parameter, provided as a value
+    :param ui_class: A Qt class to use for the UI of this :class:`.InputField`
+    :param group_by: Defines the :class:`InputField(s)<.InputField>` that controls the visibility
+        of the associated input; can be a string containing the :class:`.InputField`
+        name, a list of strings with multiple :class:`.InputField` names, or a dict
+        containing {"InputField name": condition} pairs.
+    :param group_condition: The condition for the group_by :class:`.InputField`
+        that controls the visibility of this :class:`.InputField`, provided as a value
         or a (lambda)function. If the group_by argument is provided as a
         list of strings, this argument can be either a single condition or
         a list of conditions. If the group_by argument is provided as a dict
@@ -90,7 +103,7 @@ class Parameter:
         This property returns a list of data to help formatting a command line
         interface interpreter, the list is composed of the following elements:
         - index 0: default value
-        - index 1: List of value to format an help string, that is either,
+        - index 1: List of value to format a help string, that is either,
         the name of the fields to be documented or a tuple with (helps_string,
         field)
         - index 2: type
@@ -98,7 +111,7 @@ class Parameter:
         return (self.default, self._help_fields, self.convert)
 
     def is_set(self):
-        """ Returns True if the Parameter value is set
+        """ Returns True if the InputField value is set
         """
         return self._value is not None
 
@@ -123,19 +136,17 @@ class Parameter:
             self.__class__.__name__, self.name, self._value, self.default)
 
 
-class IntegerParameter(Parameter):
-    """ :class:`.Parameter` sub-class that uses the integer type to
-    store the value.
+class IntegerInputField(InputField):
+    """ :class:`.InputField` subclass that uses the integer type to store the value.
 
-    :var value: The integer value of the parameter
+    :var value: The integer value of the InputField
 
-    :param name: The parameter name
-    :param units: The units of measure for the parameter
+    :param name: the InputField name
+    :param units: The units of measure for the InputField
     :param minimum: The minimum allowed value (default: -1e9)
     :param maximum: The maximum allowed value (default: 1e9)
-    :param default: The default integer value
-    :param ui_class: A Qt class to use for the UI of this parameter
-    :param step: int step size for parameter's UI spinbox. If None, spinbox will have step disabled
+    :param step: int step size for the field's UI spinbox. If None, spinbox will have step disabled.
+    :param \\**kwargs: keyword-arguments passed to :class:`.InputField`
     """
 
     def __init__(self, name, units=None, minimum=-1e9, maximum=1e9, step=None, **kwargs):
@@ -152,17 +163,17 @@ class IntegerParameter(Parameter):
             value, _, units = value.strip().partition(" ")
             if units != "" and units != self.units:
                 raise ValueError("Units included in string (%s) do not match"
-                                 "the units of the IntegerParameter (%s)" % (units, self.units))
+                                 "the units of the IntegerInputField (%s)" % (units, self.units))
 
         try:
             value = int(value)
         except ValueError:
-            raise ValueError("IntegerParameter given non-integer value of "
+            raise ValueError("IntegerInputField given non-integer value of "
                              "type '%s'" % type(value))
         if value < self.minimum:
-            raise ValueError("IntegerParameter value is below the minimum")
+            raise ValueError("IntegerInputField value is below the minimum")
         elif value > self.maximum:
-            raise ValueError("IntegerParameter value is above the maximum")
+            raise ValueError("IntegerInputField value is above the maximum")
 
         return value
 
@@ -179,15 +190,14 @@ class IntegerParameter(Parameter):
             self.__class__.__name__, self.name, self._value, self.units, self.default)
 
 
-class BooleanParameter(Parameter):
-    """ :class:`.Parameter` sub-class that uses the boolean type to
-    store the value.
+class BooleanInputField(InputField):
+    """ :class:`.InputField` subclass that uses the boolean type to store the value.
 
-    :var value: The boolean value of the parameter
+    :var value: The boolean value of the InputField
 
-    :param name: The parameter name
+    :param name: the InputField name
     :param default: The default boolean value
-    :param ui_class: A Qt class to use for the UI of this parameter
+    :param ui_class: A Qt class to use for the UI of this InputField
     """
 
     def convert(self, value):
@@ -197,31 +207,29 @@ class BooleanParameter(Parameter):
             elif value.lower() == "false":
                 value = False
             else:
-                raise ValueError("BooleanParameter given string value of '%s'" % value)
+                raise ValueError("BooleanInputField given string value of '%s'" % value)
         elif isinstance(value, (int, float)) and value in [0, 1]:
             value = bool(value)
         elif isinstance(value, bool):
             value = value
         else:
-            raise ValueError("BooleanParameter given non-boolean value of "
+            raise ValueError("BooleanInputField given non-boolean value of "
                              "type '%s'" % type(value))
         return value
 
 
-class FloatParameter(Parameter):
-    """ :class:`.Parameter` sub-class that uses the floating point
-    type to store the value.
+class FloatInputField(InputField):
+    """ :class:`.InputField` subclass that uses the floating point type to store the value.
 
-    :var value: The floating point value of the parameter
+    :var value: The floating point value of the InputField
 
-    :param name: The parameter name
-    :param units: The units of measure for the parameter
+    :param name: the InputField name
+    :param units: The units of measure for the InputField
     :param minimum: The minimum allowed value (default: -1e9)
     :param maximum: The maximum allowed value (default: 1e9)
     :param decimals: The number of decimals considered (default: 15)
-    :param default: The default floating point value
-    :param ui_class: A Qt class to use for the UI of this parameter
-    :param step: step size for parameter's UI spinbox. If None, spinbox will have step disabled
+    :param step: step size for the field's UI spinbox. If None, spinbox will have step disabled
+    :param \\**kwargs: keyword-arguments passed to :class:`.InputField`
     """
 
     def __init__(self, name, units=None, minimum=-1e9, maximum=1e9,
@@ -239,17 +247,17 @@ class FloatParameter(Parameter):
             value, _, units = value.strip().partition(" ")
             if units != "" and units != self.units:
                 raise ValueError("Units included in string (%s) do not match"
-                                 "the units of the FloatParameter (%s)" % (units, self.units))
+                                 "the units of the FloatInputField (%s)" % (units, self.units))
 
         try:
             value = float(value)
         except ValueError:
-            raise ValueError("FloatParameter given non-float value of "
+            raise ValueError("FloatInputField given non-float value of "
                              "type '%s'" % type(value))
         if value < self.minimum:
-            raise ValueError("FloatParameter value is below the minimum")
+            raise ValueError("FloatInputField value is below the minimum")
         elif value > self.maximum:
-            raise ValueError("FloatParameter value is above the maximum")
+            raise ValueError("FloatInputField value is above the maximum")
 
         return value
 
@@ -266,17 +274,15 @@ class FloatParameter(Parameter):
             self.__class__.__name__, self.name, self._value, self.units, self.default)
 
 
-class VectorParameter(Parameter):
-    """ :class:`.Parameter` sub-class that stores the value in a
-    vector format.
+class VectorInputField(InputField):
+    """ :class:`.InputField` subclass that stores the value in a vector format.
 
-    :var value: The value of the parameter as a list of floating point numbers
+    :var value: The value of the InputField as a list of floating point numbers
 
-    :param name: The parameter name
+    :param name: the InputField name
     :param length: The integer dimensions of the vector
-    :param units: The units of measure for the parameter
-    :param default: The default value
-    :param ui_class: A Qt class to use for the UI of this parameter
+    :param units: The units of measure for the InputField
+    :param \\**kwargs: keyword-arguments passed to :class:`.InputField`
     """
 
     def __init__(self, name, length=3, units=None, **kwargs):
@@ -293,23 +299,23 @@ class VectorParameter(Parameter):
 
             # Strip initial and final brackets
             if (value[0] != '[') or (value[-1] != ']'):
-                raise ValueError("VectorParameter must be passed a vector"
+                raise ValueError("VectorInputField must be passed a vector"
                                  " denoted by square brackets if initializing"
                                  " by string.")
             raw_list = value[1:-1].split(",")
         elif isinstance(value, (list, tuple)):
             raw_list = value
         else:
-            raise ValueError("VectorParameter given undesired value of "
+            raise ValueError("VectorInputField given undesired value of "
                              "type '%s'" % type(value))
         if len(raw_list) != self._length:
-            raise ValueError("VectorParameter given value of length "
+            raise ValueError("VectorInputField given value of length "
                              "%d instead of %d" % (len(raw_list), self._length))
         try:
             value = [float(ve) for ve in raw_list]
 
         except ValueError:
-            raise ValueError("VectorParameter given input '%s' that could "
+            raise ValueError("VectorInputField given input '%s' that could "
                              "not be converted to floats." % str(value))
 
         return value
@@ -329,15 +335,14 @@ class VectorParameter(Parameter):
             self.__class__.__name__, self.name, self._value, self.units, self._length)
 
 
-class ListParameter(Parameter):
-    """ :class:`.Parameter` sub-class that stores the value as a list.
+class ListInputField(InputField):
+    """ :class:`.InputField` subclass that stores the value as a list.
     String representation of choices must be unique.
 
-    :param name: The parameter name
+    :param name: the InputField name
     :param choices: An explicit list of choices, which is disregarded if None
-    :param units: The units of measure for the parameter
-    :param default: The default value
-    :param ui_class: A Qt class to use for the UI of this parameter
+    :param units: The units of measure for the InputField
+    :param \\**kwargs: keyword-arguments passed to :class:`.InputField`
     """
 
     def __init__(self, name, choices=None, units=None, **kwargs):
@@ -356,7 +361,7 @@ class ListParameter(Parameter):
 
     def convert(self, value):
         if self._choices is None:
-            raise ValueError("ListParameter cannot be set since "
+            raise ValueError("ListInputField cannot be set since "
                              "allowed choices are set to None.")
 
         # strip units if included
@@ -378,17 +383,16 @@ class ListParameter(Parameter):
         return tuple(self._choices.values())
 
 
-class PhysicalParameter(VectorParameter):
-    """ :class:`.VectorParameter` sub-class of 2 dimensions to store a value
+class PhysicalInputField(VectorInputField):
+    """ :class:`.VectorInputField` subclass of 2 dimensions to store a value
     and its uncertainty.
 
-    :var value: The value of the parameter as a list of 2 floating point numbers
+    :var value: The value of the InputField as a list of 2 floating point numbers
 
-    :param name: The parameter name
+    :param name: the InputField name
     :param uncertainty_type: Type of uncertainty, 'absolute', 'relative' or 'percentage'
-    :param units: The units of measure for the parameter
-    :param default: The default value
-    :param ui_class: A Qt class to use for the UI of this parameter
+    :param units: The units of measure for the InputField
+    :param \\**kwargs: keyword-arguments passed to :class:`.InputField`
     """
 
     def __init__(self, name, uncertaintyType='absolute', **kwargs):
@@ -406,22 +410,22 @@ class PhysicalParameter(VectorParameter):
 
             # Strip initial and final brackets
             if (value[0] != '[') or (value[-1] != ']'):
-                raise ValueError("VectorParameter must be passed a vector"
+                raise ValueError("PhysicalInputField must be passed a vector"
                                  " denoted by square brackets if initializing"
                                  " by string.")
             raw_list = value[1:-1].split(",")
         elif isinstance(value, (list, tuple)):
             raw_list = value
         else:
-            raise ValueError("VectorParameter given undesired value of "
+            raise ValueError("PhysicalInputField given undesired value of "
                              "type '%s'" % type(value))
         if len(raw_list) != self._length:
-            raise ValueError("VectorParameter given value of length "
+            raise ValueError("PhysicalInputField given value of length "
                              "%d instead of %d" % (len(raw_list), self._length))
         try:
             value = [float(ve) for ve in raw_list]
         except ValueError:
-            raise ValueError("VectorParameter given input '%s' that could "
+            raise ValueError("PhysicalInputField given input '%s' that could "
                              "not be converted to floats." % str(value))
         # Uncertainty must be non-negative
         value[1] = abs(value[1])
@@ -468,6 +472,229 @@ class PhysicalParameter(VectorParameter):
             self.__class__.__name__, self.name, self._value, self.units, self._utype.value)
 
 
+class Parameter(InputField):
+    """ Encapsulates the information for a parameter of an experiment with information about the
+    name.
+
+    The data dat is stored in a parameter is written to the parameters section of the data file
+    when an experiment is queued.
+    Information that is stored as a parameter is loaded to the inputs-widget of a ManagedWindow by
+    using the `Use These Parameters` option.
+
+    :var value: The value of the :class:`.Parameter`
+
+    :param name: The :class:`.Parameter` name
+    :param default: The default value
+    :param ui_class: A Qt class to use for the UI of this :class:`.Parameter`
+    :param group_by: Defines the :class:`InputField(s)<.InputField>` that controls the visibility
+        of the associated input; can be a string containing the :class:`.Parameter`
+        name, a list of strings with multiple :class:`.InputField` names, or a dict
+        containing {"InputField name": condition} pairs.
+    :param group_condition: The condition for the group_by :class:`.InputField`
+        that controls the visibility of this :class:`.Parameter`, provided as a value
+        or a (lambda)function. If the group_by argument is provided as a
+        list of strings, this argument can be either a single condition or
+        a list of conditions. If the group_by argument is provided as a dict
+        this argument is ignored.
+    """
+    pass
+
+
+class IntegerParameter(Parameter, IntegerInputField):
+    """ :class:`.Parameter` subclass that uses the integer type to store the value..
+
+    :var value: The integer value of the Parameter
+
+    :param name: the Parameter name
+    :param units: The units of measure for the Parameter
+    :param minimum: The minimum allowed value (default: -1e9)
+    :param maximum: The maximum allowed value (default: 1e9)
+    :param step: int step size for the field's UI spinbox. If None, spinbox will have step disabled.
+    :param \\**kwargs: keyword-arguments passed to :class:`.Parameter`
+    """
+    pass
+
+
+class BooleanParameter(Parameter, BooleanInputField):
+    """ :class:`.Parameter` subclass that uses the boolean type to store the value.
+
+    :var value: The boolean value of the parameter
+
+    Refer to :class:`.Parameter` for the initialization arguments and keyword-arguments.
+    """
+    pass
+
+
+class FloatParameter(Parameter, FloatInputField):
+    """ :class:`.Parameter` subclass that uses the floating point type to store the value.
+
+    :var value: The floating point value of the parameter
+
+    :param name: the parameter name
+    :param units: The units of measure for the parameter
+    :param minimum: The minimum allowed value (default: -1e9)
+    :param maximum: The maximum allowed value (default: 1e9)
+    :param decimals: The number of decimals considered (default: 15)
+    :param step: step size for the field's UI spinbox. If None, spinbox will have step disabled
+    :param \\**kwargs: keyword-arguments passed to :class:`.Parameter`
+    """
+    pass
+
+
+class VectorParameter(Parameter, VectorInputField):
+    """ :class:`.Parameter` subclass that stores the value in a vector format.
+
+    :var value: The value of the parameter as a list of floating point numbers
+
+    :param name: the parameter name
+    :param length: The integer dimensions of the vector
+    :param units: The units of measure for the parameter
+    :param \\**kwargs: keyword-arguments passed to :class:`.Parameter`
+    """
+    pass
+
+
+class ListParameter(Parameter, ListInputField):
+    """ :class:`.Parameter` subclass that stores the value as a list.
+    String representation of choices must be unique.
+
+    :param name: the parameter name
+    :param choices: An explicit list of choices, which is disregarded if None
+    :param units: The units of measure for the parameter
+    :param \\**kwargs: keyword-arguments passed to :class:`.Parameter`
+    """
+    pass
+
+
+class PhysicalParameter(Parameter, PhysicalInputField):
+    """ :class:`.VectorParameter` subclass of 2 dimensions to store a value
+    and its uncertainty.
+
+    :var value: The value of the parameter as a list of 2 floating point numbers
+
+    :param name: the parameter name
+    :param uncertainty_type: Type of uncertainty, 'absolute', 'relative' or 'percentage'
+    :param units: The units of measure for the parameter
+    :param \\**kwargs: keyword-arguments passed to :class:`.Parameter`
+    """
+    pass
+
+
+class Metadata(InputField):
+    """ Encapsulates the information for metadata of the experiment with
+    information about the name, the fget function and the units, if supplied.
+    If no fget function is specified, the value property will return the
+    latest set value of the parameter (or default if never set).
+    Subclass of :class:`.InputField`
+
+    :var value: The value of the parameter. This returns (if a value is set)
+        the value obtained from the `fget` (after evaluation) or a manually
+        set value. Returns `None` if no value has been set
+
+    :param name: The metadata name
+    :param fget: The metadata fget function; can be provided as a callable,
+        or as a string, in which case it is assumed to be the name of a
+        method or attribute of the `Procedure` class in which the Metadata is
+        defined. Passing a string also allows for nested attributes by separating
+        them with a period (e.g. to access an attribute or method of an
+        instrument) where only the last attribute can be a method.
+    :param units: The metadata units
+
+        .. deprecated:: 0.14
+            Removed from the base :class:`Metadata` class; the numeric subclasses
+            :class:`.IntegerMetadata`, :class:`.FloatMetadata`, :class:`.VectorMetadata`,
+            :class:`.ListMetadata`, and :class:`.PhysicalMetadata` do provide the possibility to
+            define units.
+
+    :param default: The default value, in case no value is assigned or if no
+        fget method is provided
+    :param fmt: A string used to format the value upon writing it to a file.
+        Default is "%s"
+
+        .. deprecated:: 0.14
+            The formatting is now defined by the specific :class:`.Metadata` subclasses; the
+            :class:`.Metadata` base class uses :code:`str` to convert the value to a string.
+
+    """
+
+    def __init__(self, name, fget=None, units=None, fmt=None, **kwargs):
+        self.fget = fget
+        self.evaluated = False
+        super().__init__(name, **kwargs)
+
+        if units is not None:
+            self.units = units
+            warn("units parameter is deprecated, use appropriate Metadata subclass instead"
+                 " instead", FutureWarning)
+
+        if fmt is not None:
+            self.fmt = fmt
+            warn("fmt parameter is deprecated, use appropriate Metadata subclass instead"
+                 " instead", FutureWarning)
+
+    def evaluate(self, parent=None, new_value=None):
+        if new_value is not None and self.fget is not None:
+            raise ValueError("Metadata with a defined fget method"
+                             " cannot be manually assigned a value")
+        elif new_value is not None:
+            self._value = new_value
+        elif self.fget is not None:
+            self._value = self.eval_fget(parent)
+
+        self.evaluated = True
+        return self.value
+
+    def eval_fget(self, parent):
+        fget = self.fget
+        if isinstance(fget, str):
+            obj = parent
+            for obj_name in fget.split('.'):
+                obj = getattr(obj, obj_name)
+            fget = obj
+
+        if callable(fget):
+            return fget()
+        else:
+            return fget
+
+    def __str__(self):
+        # Deprecated, can be removed once the units and fmt parameters are removed from the
+        # Metadata base class
+        if hasattr(self, 'fmt'):
+            result = self.fmt % self.value
+        else:
+            result = super().__str__()
+
+        if hasattr(self, 'units') and self.units is not None:
+            result += " %s" % self.units
+
+        return result
+
+
+class IntegerMetadata(Metadata, IntegerInputField):
+    pass
+
+
+class BooleanMetadata(Metadata, BooleanInputField):
+    pass
+
+
+class FloatMetadata(Metadata, FloatInputField):
+    pass
+
+
+class VectorMetadata(Metadata, VectorInputField):
+    pass
+
+
+class ListMetadata(Metadata, ListInputField):
+    pass
+
+
+class PhysicalMetadata(Metadata, PhysicalInputField):
+    pass
+
+
 class Measurable:
     """ Encapsulates the information for a measurable experiment parameter
     with information about the name, fget function and units if supplied.
@@ -507,83 +734,3 @@ class Measurable:
     @value.setter
     def value(self, value):
         self._value = value
-
-
-class Metadata(object):
-    """ Encapsulates the information for metadata of the experiment with
-    information about the name, the fget function and the units, if supplied.
-    If no fget function is specified, the value property will return the
-    latest set value of the parameter (or default if never set).
-
-    :var value: The value of the parameter. This returns (if a value is set)
-        the value obtained from the `fget` (after evaluation) or a manually
-        set value. Returns `None` if no value has been set
-
-    :param name: The parameter name
-    :param fget: The parameter fget function; can be provided as a callable,
-        or as a string, in which case it is assumed to be the name of a
-        method or attribute of the `Procedure` class in which the Metadata is
-        defined. Passing a string also allows for nested attributes by separating
-        them with a period (e.g. to access an attribute or method of an
-        instrument) where only the last attribute can be a method.
-    :param units: The parameter units
-    :param default: The default value, in case no value is assigned or if no
-        fget method is provided
-    :param fmt: A string used to format the value upon writing it to a file.
-        Default is "%s"
-
-    """
-    def __init__(self, name, fget=None, units=None, default=None, fmt="%s"):
-        self.name = name
-        self.units = units
-
-        self._value = default
-        self.fget = fget
-        self.fmt = fmt
-
-        self.evaluated = False
-
-    @property
-    def value(self):
-        if self.is_set():
-            return self._value
-        else:
-            raise ValueError("Metadata value is not set")
-
-    def is_set(self):
-        """ Returns True if the Parameter value is set
-        """
-        return self._value is not None
-
-    def evaluate(self, parent=None, new_value=None):
-        if new_value is not None and self.fget is not None:
-            raise ValueError("Metadata with a defined fget method"
-                             " cannot be manually assigned a value")
-        elif new_value is not None:
-            self._value = new_value
-        elif self.fget is not None:
-            self._value = self.eval_fget(parent)
-
-        self.evaluated = True
-        return self.value
-
-    def eval_fget(self, parent):
-        fget = self.fget
-        if isinstance(fget, str):
-            obj = parent
-            for obj_name in fget.split('.'):
-                obj = getattr(obj, obj_name)
-            fget = obj
-
-        if callable(fget):
-            return fget()
-        else:
-            return fget
-
-    def __str__(self):
-        result = self.fmt % self.value
-
-        if self.units is not None:
-            result += " %s" % self.units
-
-        return result